pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "TkEasyGUI"
version = "1.0.0"
dependencies = [
  "Pillow",
  "pyperclip",
]
requires-python = ">=3.9"
authors = [
  { name="kujirahand", email="web@kujirahand.com" },
]
maintainers = [
 { name="kujirahand", email="web@kujirahand.com" },
]
description = "TkEasyGUI is simple GUI Library for Python3 with Tkinter"
readme = "README.md"
license = {file = "LICENSE"}
keywords = ["GUI", "Tkinter", "PySimpleGUI"]
classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Software Development :: User Interfaces",
]

[project.urls]
Homepage = "https://github.com/kujirahand/tkeasygui-python"
Documentation = "https://github.com/kujirahand/tkeasygui-python/blob/main/README.md"
Repository = "https://github.com/kujirahand/tkeasygui-python/"
Issues = "https://github.com/kujirahand/tkeasygui-python/issues"
Changelog = "https://github.com/kujirahand/tkeasygui-python/releases"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]
